Add move() to Piece interface
Added a move function to Piece.ts.
Changed has_moved to moved and added it to Piece interface as opposed to just the Pawn object.
Board now rotates depending on whose go it is.
Player objects now keep track of taken pieces, which are appended to when a piece is taken.